[core] Increase admin possibilities
This commit adds more possibilities for the admin to control this MOD. He can now
access a new token {REPORTER} (just the username of the reporter) {REPORTER_FULL}
(the full profile link to the profile of the reporter). You can now also set the
topic title that is used based upon its a pm or post report. Tokens can be used
in the titles as well.
Also some overall cleaning in the acp template file.