[feature/forum-designate] Begin the forum designate feature.
Some initial work on the forum designate feature. This feature allows an administrator
to decided which forum will be used for the report posts on a forum basis. If none is
selected the default destination forum will be used, this default forum can be set
through the MODs own ACP page.